En teoria ya esta ready pa la accion

+++ src/handlers/empleado.ts
@@ -1,11 +1,11 @@
 import { prisma } from "../prisma/client";
 import { Request } from "express";
 import * as argon2 from "argon2";
+import jwt from "jsonwebtoken";
 
 export module HandleEmpleado {
   export async function crear(req: Request) {
     let json = req.body;
-
     if (Array.isArray(json)) {
       for (const data of json) {
         if (!data["password_empleado"]) {
@@ -39,13 +39,20 @@ export module HandleEmpleado {
       throw new Error("No pusiste correo_empleado en la direccion papu :v");
     }
 
+    let token = req.params.token;
+    let decoded = jwt.verify(token, "secret")
+
+    if (!decoded) {
+      throw new Error("Token invalido");
+    }
+
     await prisma.creador.deleteMany({
       where: {
         correo_creador: req.params["correo"],
       },
     });
 
-    await prisma.modelo.deleteMany({
+    await prisma.proyecto.deleteMany({
       where: {
         correo_creador: req.params["correo"],
       },
